Moving files and tests to tests and usage tests to sample.py
